:recycle: begin migrating functions from mongo driver and learning object interactor
