+++ src/LearningObjectSearch/drivers/index.ts
@@ -1,3 +1,3 @@
-import { ElasticSearchLearningObjectDatastore } from './LearningObjectDatastore/ElasticSearchLearningObjectDatastore';
+import { ElasticSearchLearningObjectDatastore } from './LearningObjectDatastore/ElasticSearchLearningObjectDatastore/ElasticSearchLearningObjectDatastore';
 
 export { ElasticSearchLearningObjectDatastore };
